+++ src/main/java/net/brnbrd/delightful/compat/AquaticCompat.java
@@ -1,0 +1,10 @@
+package net.brnbrd.delightful.compat;
+
+import net.minecraft.world.food.FoodProperties;
+import vectorwing.farmersdelight.common.FoodValues;
+
+public class AquaticCompat {
+	public static final String mulberry_pie = "mulberry_pie";
+
+	public static final FoodProperties MULBERRY_PIE_SLICE = FoodValues.PIE_SLICE;
+}