+++ src/main/java/net/brnbrd/delightful/data/gen/DelightfulBlockTagProvider.java
@@ -47,6 +47,7 @@ public class DelightfulBlockTagProvider extends BlockTagsProvider {
 				.add(DelightfulBlocks.SOURCE_BERRY_PIE.get())
 				.add(DelightfulBlocks.GLOOMGOURD_PIE.get())
 				.add(DelightfulBlocks.CHORUS_PIE.get())
+				.add(DelightfulBlocks.MULBERRY_PIE.get())
 				.add(DelightfulBlocks.BAKLAVA.get())
 				.addOptional(Util.rl(Mods.UG, "gloomgourd"))
 				.addOptional(Util.rl(Mods.UG, "carved_gloomgourd"));