+++ changelog.md
@@ -6,6 +6,8 @@
 ### Updates
 - Matcha products (but not Matcha itself) now give level 2 of the green tea effect
 - Gummy effects are now displayed in their tooltip
+- Placeable Mulberry Pie, added Mulberry Pie Slice (for [Upgrade Aquatic](https://www.curseforge.com/minecraft/mc-mods/upgrade-aquatic))
+  - Retexture of Mulberry Pie in `Delightful Overhauls` built-in resource pack
 
 
 # 3.7